chore: moving server starting into webserver dir
