+++ package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "src/frameworks/server.js",
   "scripts": {
-    "start": "node src/frameworks/server.js",
-    "dev": "nodemon --ignore test/ -L src/frameworks/server.js",
+    "start": "node src/frameworks/webserver/server.js",
+    "dev": "nodemon --ignore test/ -L src/frameworks/webserver/server.js",
     "test": "cross-env NODE_ENV=test mocha 'test/**/*.js' --runInBand --exit",
     "test:watch": "cross-env NODE_ENV=test mocha 'test/**/*.js' --watch --runInBand --exit",
     "test:coverage": "nyc --reporter=text --reporter=html cross-env NODE_ENV=test mocha 'test/**/*.js' --exit",
