+++ src/frameworks/webserver/server.js
@@ -1,6 +1,6 @@
 const app = require('./app');
 
-const { NODE_ENV, APP_PORT } = require('./config');
+const { NODE_ENV, APP_PORT } = require('../config');
 
 if (NODE_ENV !== 'test') {
   app.listen(APP_PORT, () => {
