+++ test/app.spec.js
@@ -4,7 +4,7 @@ const request = require('supertest');
 const { expect } = chai;
 
 describe('Main App', () => {
-  let app = require('../src/frameworks/app');
+  let app = require('../src/frameworks/webserver/app');
 
   it('should respond successfully to /', async () => {
     const res = await request(app).get('/').expect(200);
